% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_denovo.R
\name{csHeatmap}
\alias{csHeatmap}
\title{Cosine Similarity Matrix Heatmap}
\usage{
csHeatmap(
  z,
  pal = NA,
  zmin = 0,
  zmax = 1,
  cex = 1,
  ddx = 0.03,
  ddy = 0.03,
  mar = c(1, 1, 5, 2),
  grid.col = "gray",
  na.col = "gray",
  legend = NA,
  rescale = FALSE,
  lwd = 0.5,
  col.break = 30,
  show.lsap = TRUE
)
}
\arguments{
\item{show.lsap}{Display LSAP assignment as rectangles}

\item{cs}{Cosine similarity matrix}
}
\description{
Cosine Similarity Matrix Heatmap
}
\examples{
set.seed(135)
}
