% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\docType{class}
\name{DeepSig-class}
\alias{DeepSig-class}
\title{Class \code{DeepSig} for storing catalog and exposure}
\value{
Object of class \code{DeepSig}
}
\description{
\code{S4} class with storage for input catalog, reference signatures,
and output exposure inferred.
}
\section{Slots}{

\describe{
\item{\code{catalog}}{Matrix of mutation count with samples in columns
and mutation types in rows}

\item{\code{signat}}{Matrix of reference signatures; each column
stores individual signatures proportions with mutation
types in rows}

\item{\code{tmb}}{Vector of total mutation loads for each sample}

\item{\code{expos}}{Matrix of inferred proportions of signatures
exposure; each row gives proportions of each
sample with signatures in columns}

\item{\code{pvalue}}{P-values of exposures estimated from permutation tests;
same dimension as \code{expos}}

\item{\code{logLik}}{Log likelihood}

\item{\code{misc}}{List of factorization results and measures for de novo runs}
}}

