% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepSig-package.R
\docType{package}
\name{DeepSig-package}
\alias{DeepSig-package}
\title{DeepSig: Deep Learning-Based Mutational Signature Inference for WES and MSK-IMPACT data}
\description{
Infers mutation signature exposures from mutation catalog data using cancer-specific and pan-cancer reference signature lists via deep learning-trained models. Also includes de novo extraction routines. Suitable for analysis of catalogs derived from limited mutation count data such as whole exome sequencing or MSK-IMPACT.
}
\keyword{internal}
