% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R, R/DeepSig.R
\docType{package}
\name{DeepSig}
\alias{DeepSig}
\title{Create \code{DeepSig} object}
\usage{
DeepSig(data, signat = "v2")
}
\arguments{
\item{data}{Input data of mutation catalog; samples in columns
and mutation types in rows.}

\item{signat}{Reference signatures table; default is COSMIC
version 3 signatures. It can be a matrix of the table, the file name
containing the table, or character code \code{c('v2', 'SA','SP')},
signifying \code{COSMIC v2}, \code{sigAnalyzer}, or
\code{SigProfiler} references (the latter two are COSMIC v3).}
}
\value{
Object of class \code{DeepSig}.
}
\examples{
set.seed(130)
data <- rmultinom(n = 5, size = 100, prob = rep(1,96))
rownames(data) <- trinucleotides()
colnames(data) <- paste0('S',seq(5))
s <- DeepSig(data)
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt',
                  package='DeepSig'), header=TRUE, sep='\t')
b <- DeepSig(data)
}
