% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_denovo.R
\name{denovoCosim}
\alias{denovoCosim}
\title{Generate Cosine Similarity Matrix Between De novo and Reference Signatures}
\usage{
denovoCosim(sig, ref.sig = "cosmic")
}
\arguments{
\item{sig}{Input de novo signature matrix}

\item{ref.sig}{Reference signature set}
}
\value{
Cosine similarity matrix
}
\description{
Generate Cosine Similarity Matrix Between De novo and Reference Signatures
}
\examples{
set.seed(135)
}
