% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_denovo.R
\name{denovoProfiles}
\alias{denovoProfiles}
\title{Plot De Novo Signatures and Closest Reference Signatures}
\usage{
denovoProfiles(object, ref.sig = "cosmic")
}
\arguments{
\item{object}{Object whose signature set is to be plotted}

\item{ref.sig}{Reference signature set}
}
\description{
Plot De Novo Signatures and Closest Reference Signatures
}
