% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{expos,DeepSig-method}
\alias{expos,DeepSig-method}
\title{Accessor for exposure}
\usage{
\S4method{expos}{DeepSig}(object)
}
\arguments{
\item{object}{Object containing \code{expo}}
}
\value{
\code{expos} data frame
}
\description{
Accessor for exposure
}
