% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractSig}
\alias{extractSig}
\title{Infer Signature Proportions}
\usage{
extractSig(
  object,
  method = "mle",
  screen = NA,
  itmax = 1000,
  tol = 1e-04,
  min.tmb = 1,
  compute.pval = FALSE,
  nperm = 1000,
  progress.bar = FALSE,
  pvtest = "permutation",
  cosmic = FALSE,
  Kmin = 2,
  K = 2,
  alpha = 1,
  Kmax = 30,
  nrun = 10,
  useC = TRUE,
  initializer = "random",
  nboot = 0,
  verbose = 2,
  fix.h = NA,
  s2a = NA,
  sig.collapse = NA,
  ...
)
}
\arguments{
\item{object}{Object of class \code{DeepSig}}

\item{method}{Refitting method; \code{mle} for maximum likelihood (default) or
\code{mutCone} for mutationalCone.}

\item{screen}{Logical matrix of dimension \code{nsample \times nsignature} whose
rows indicate signatures to be included in refitting for samples}

\item{itmax}{Maximum number of iterations for maximum likelihood estimate}

\item{tol}{Tolerance for convergence}

\item{min.tmb}{Minimum number of mutations in each sample. If \code{tmb} is less,
\code{NA} will be returned for the sample.}

\item{compute.pval}{Estimate p-values}

\item{nperm}{Number of permutations}

\item{progress.bar}{Display progress bar}

\item{pvtest}{Algorithm for p-value computation; 
\code{c('permutation','lrt','x.permutation')} for permutation resampling of 
signatures, likelihood ratio test (asymptotic formula), or
permutation of count data.}
}
\description{
Use known signature list to find the most likely exposures in samples
}
\examples{
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt', package='DeepSig'))
b <- DeepSig(data)
b <- extractSig(b, progress.bar = TRUE)
b_pv <- extractSig(b, compute.pval = TRUE, progress.bar = TRUE)
}
