% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{misc,DeepSig-method}
\alias{misc,DeepSig-method}
\title{Accessor for misc}
\usage{
\S4method{misc}{DeepSig}(object)
}
\arguments{
\item{object}{Object containing \code{misc}}
}
\value{
List \code{misc}
}
\description{
Accessor for misc
}
