% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFetch.R
\name{modelFetch}
\alias{modelFetch}
\title{Load Pre-trained Models from Github}
\usage{
modelFetch(
  url = "https://api.github.com",
  path = "/repos/mskcc/DeepSig/contents/inst/extdata/dlsig/",
  version = "v0.95",
  cancer.type = "breast",
  token = NA,
  verbose = TRUE,
  model.path = NA
)
}
\arguments{
\item{url}{Main API url}

\item{path}{Repository path of the model}

\item{version}{Version number}

\item{cancer.type}{Cancer type}

\item{token}{Authentication token; if `NA`, access attempted without one}

\item{verbose}{Verbosity level}

\item{model.path}{Directory path where the models will be downloaded; if `NA`,
`tempdir()` will be used}

\item{min.M}{Minimum no. of mutations}

\item{min.attr}{Mininum attribution}
}
\description{
Use REST API of the github and download pre-trained models onto the 
local machine
}
\examples{
dir <- modelFetch()
print(dir)

}
