% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotree.R
\name{oncoTree}
\alias{oncoTree}
\title{Convert OncoTree Code into Cancer Type}
\usage{
oncoTree(onco)
}
\arguments{
\item{oncotree}{Oncotree code}
}
\description{
Convert OncoTree Code into Cancer Type
}
\examples{
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt',package='DeepSig'))
z <- DL.call(catalog = t(data), cancer.type = 'nsclc', alpha = 0.05)
head(z$exposure.fltrd)

}
