% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgnmf.R
\name{pgnmf}
\alias{pgnmf}
\title{piggyback NMF}
\usage{
pgnmf(
  x,
  stock = NULL,
  method = "bnmf",
  filter = FALSE,
  cutoff = NULL,
  discrete = FALSE,
  progress.bar = TRUE,
  nrun = 1,
  verbose = 1,
  alpha = 1,
  nboot = 0,
  pg.dev = TRUE,
  pg.pval = FALSE,
  xtype = NA,
  prefilter = NA,
  tspec = FALSE,
  s2a = NA,
  ...
)
}
\arguments{
\item{x}{Catalog matrix to be analyzed}

\item{stock}{\code{DeepSig} object containing the stock solution}

\item{method}{\code{'nmf'} or \code{'bnmf'} for inference engine}

\item{filter}{Filter exposure using CV cutoff \code{cutoff}}

\item{cutoff}{CV cutoff matrix for filtering}

\item{discrete}{Convert proportions into multinomial counts}

\item{alpha}{Dirichlet alpha for initial guess of exposure vector}

\item{nboot}{No. of bootstrapping runs}

\item{pg.pval}{Estimate p-values by permutation}

\item{prefilter}{List of allowed cancer type to be used for prefiltering; 
if \code{NA}, skip prefiltering}

\item{pg.lr}{Compute deviance for each signature}
}
\value{
List of \code{mean} and coefficient of variation \code{cv} (sd/mean) of exposure matrices
}
\description{
Enables de novo inference of single sample using pre-inferred exposure 
and signature sets for a large data set (stock)
}
