% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities3.R
\name{reorderSig}
\alias{reorderSig}
\title{Rearrange Signatures in Object}
\usage{
reorderSig(object, order, label = NULL)
}
\arguments{
\item{object}{\code{DeepSig} object}

\item{order}{Reordering index of signatures}

\item{label}{New signature labels}
}
\description{
Rearrange Signatures in Object
}
