% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runBootstrap}
\alias{runBootstrap}
\title{Run multiple decomposition with bootstrapping}
\usage{
runBootstrap(
  catalog,
  method = "nmf",
  verbose = 2,
  nboot = 100,
  K = 2,
  nrun = 100,
  progress.bar = TRUE,
  Kmin = 2,
  Kmax = 2,
  reorder = TRUE,
  ...
)
}
\description{
Run multiple decomposition with bootstrapping
}
