% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{show,DeepSig-method}
\alias{show,DeepSig-method}
\title{Display \code{DeepSig} object}
\usage{
\S4method{show}{DeepSig}(object)
}
\arguments{
\item{object}{Object of class \code{DeepSig}}
}
\value{
\code{NULL}
}
\description{
Display the class and dimension
}
