% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_denovo.R
\name{sigLSAP}
\alias{sigLSAP}
\title{Automatic Signature Annotation}
\usage{
sigLSAP(x)
}
\arguments{
\item{x}{Signature cosine similarity matrix (rows: de novo, columns = reference)}
}
\value{
Vector of assigned reference signature names for de novo signatures
}
\description{
Automatic Signature Annotation
}
