% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{signat,DeepSig-method}
\alias{signat,DeepSig-method}
\title{Accessor for signature}
\usage{
\S4method{signat}{DeepSig}(object)
}
\arguments{
\item{object}{Object containing signatures}
}
\value{
\code{signat} data frame
}
\description{
Accessor for signature
}
