% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities3.R
\name{sigplot}
\alias{sigplot}
\title{Plot signature profile}
\usage{
sigplot(x, error = NULL, lwd = 0.5, mgp = c(2, 0.3, 0))
}
\arguments{
\item{x}{Signature profile vector; must be of length 96 and named for nucleotide contexts}

\item{lwd}{Line thickness for error bars}
}
\value{
Return vector of \code{barplot}
}
\description{
Plot signature profile
}
