% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{simulateCatalog}
\alias{simulateCatalog}
\title{Generate simulated mutational catalog}
\usage{
simulateCatalog(W, n, min.M, max.M)
}
\arguments{
\item{W}{Signature matrix with rows 96 contexts and columns signatures}

\item{n}{Number of catalogs to generate}

\item{min.M}{Minimum mutation count}

\item{max.M}{Maximum mutation count}
}
\value{
Catalog matrix
}
\description{
Generate simulated mutational catalog
}
