% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{tmb,DeepSig-method}
\alias{tmb,DeepSig-method}
\title{Accessor for TMB}
\usage{
\S4method{tmb}{DeepSig}(object)
}
\arguments{
\item{object}{Object containing \code{tmb}}
}
\value{
\code{tmb} data frame
}
\description{
Accessor for TMB
}
