% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL.R
\name{DL.call}
\alias{DL.call}
\title{Use DL-trained Model to Make Signature Calls}
\usage{
DL.call(
  catalog,
  cancer.type = "pancancer",
  model.path = "./.DeepSig",
  mode = "catalog",
  verbose = 1,
  progress.bar = TRUE,
  min.M = 1,
  min.attr = 1,
  ...
)
}
\arguments{
\item{catalog}{Matrix of mutational catalog with samples in rows and 96 trinucleotides 
in columns}

\item{cancer.type}{Cancer type}

\item{model.path}{Directory path for trained models, If `NA`, defaults to
model corresponding to `cancer.type` and will query github API to download}

\item{verbose}{Verbosity level}

\item{progress.bar}{Progress bar}

\item{min.M}{Minimum no. of mutations}

\item{min.attr}{Mininum attribution}

\item{...}{Other parameters to `modelFetch`}

\item{ref.sig}{File path for reference signatures. If `NA`, defaults
to `cancer.type`}

\item{threshold}{File path for thresholds. If `NA`, defaults to `cancer.type`}
}
\description{
Use DL-trained Model to Make Signature Calls
}
\examples{
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt',package='DeepSig'))
z <- DL.call(catalog = t(data), cancer.type = 'braest')
head(z$exposure.fltrd)

}
