% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL.R
\name{DL.call}
\alias{DL.call}
\title{Use DL-trained Model to predict HRD}
\usage{
DL.call(
  catalog,
  cancer.type = "pan_cancer",
  model.path = NA,
  ref.sig = NA,
  threshold = NA,
  verbose = 1,
  mbins = NA,
  progress.bar = TRUE,
  min.M = 1,
  min.attr = 1,
  min.attr.pole = 10,
  alpha = NA
)
}
\arguments{
\item{catalog}{Matrix of mutational catalog with samples in rows and 96 trinucleotides 
in columns}

\item{cancer.type}{Cancer type}

\item{model.path}{Directory path for trained models}

\item{ref.sig}{File path for reference signatures}

\item{threshold}{File path for thresholds}

\item{verbose}{Verbosity level}

\item{progress.bar}{Progress bar}

\item{min.M}{Minimum no. of mutations}

\item{min.attr}{Mininum attribution}

\item{min.attr.pole}{Minimum attribution for POLE signatures}

\item{alpha}{False positive rate}
}
\description{
Use DL-trained Model to predict HRD
}
\examples{
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt',package='DeepSig'))
z <- DL.call(catalog = t(data), cancer.type = 'nsclc', alpha = 0.05)
head(z$exposure.fltrd)

}
