% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL.R
\name{DL.call}
\alias{DL.call}
\title{Use DL-trained Model to predict HRD}
\usage{
DL.call(
  catalog,
  cancer.type = "breast",
  model.path = NA,
  ref.sig = NA,
  threshold = NA,
  verbose = 1,
  mbins = NA,
  progress.bar = TRUE,
  min.attr = 1,
  min.attr.pole = 10,
  alpha = NA
)
}
\arguments{
\item{catalog}{Matrix of mutational catalog with samples in rows and 96 trinucleotides 
in columns}

\item{cancer.type}{Cancer type}

\item{model.path}{Directory path for trained models}

\item{ref.sig}{File path for reference signatures}

\item{threshold}{File path for thresholds}

\item{verbose}{Verbosity level}

\item{progress.bar}{Progress bar}
}
\description{
Use DL-trained Model to predict HRD
}
